+++ Vector3.java
@@ -1,6 +1,10 @@
-/**
- * Created by Joshua on 3/6/2015.
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
  */
+package nextgen;
+
 public class Vector3
 {
     public double x, y, z;
